removed unsupported chain and update system parachain
